Rename access_key.tsx to AccessKey.tsx.
